Add files via upload
This program uses the switch to turn the program functionality on and off. When in resting state, if the button is pressed, the program outputs a mean sentence. If you push the joystick left and then push the button, the program outputs a confused sentence. If you push the joystick down and click the button, the program enters a third state and outputs a nice sentence.